fix: [CCM-12664]: Split PerspectivesQuery into PerspectivesQuery and PerspectiveService
