+++ 374-ceng-graphql/src/main/java/io/harness/ccm/graphql/query/perspectives/PerspectivesQuery.java
@@ -8,14 +8,10 @@
 package io.harness.ccm.graphql.query.perspectives;
 
 import static io.harness.annotations.dev.HarnessTeam.CE;
-import static io.harness.ccm.commons.constants.ViewFieldConstants.AWS_ACCOUNT_FIELD;
 import static io.harness.ccm.rbac.CCMRbacPermissions.PERSPECTIVE_VIEW;
 
 import io.harness.annotations.dev.OwnedBy;
-import io.harness.ccm.graphql.core.perspectives.PerspectiveFieldsHelper;
-import io.harness.ccm.graphql.core.perspectives.PerspectiveOverviewStatsHelper;
-import io.harness.ccm.graphql.core.perspectives.PerspectiveTimeSeriesHelper;
-import io.harness.ccm.graphql.dto.common.StatsInfo;
+import io.harness.ccm.graphql.core.perspectives.PerspectiveService;
 import io.harness.ccm.graphql.dto.perspectives.PerspectiveData;
 import io.harness.ccm.graphql.dto.perspectives.PerspectiveEntityStatsData;
 import io.harness.ccm.graphql.dto.perspectives.PerspectiveFieldsData;
@@ -26,33 +22,21 @@ import io.harness.ccm.graphql.utils.GraphQLUtils;
 import io.harness.ccm.graphql.utils.annotations.GraphQLApi;
 import io.harness.ccm.rbac.CCMRbacHelper;
 import io.harness.ccm.views.dto.PerspectiveTimeSeriesData;
-import io.harness.ccm.views.entities.ViewQueryParams;
 import io.harness.ccm.views.graphql.QLCEView;
 import io.harness.ccm.views.graphql.QLCEViewAggregation;
 import io.harness.ccm.views.graphql.QLCEViewFilterWrapper;
 import io.harness.ccm.views.graphql.QLCEViewGroupBy;
 import io.harness.ccm.views.graphql.QLCEViewPreferences;
 import io.harness.ccm.views.graphql.QLCEViewSortCriteria;
-import io.harness.ccm.views.graphql.QLCEViewTrendData;
-import io.harness.ccm.views.graphql.QLCEViewTrendInfo;
-import io.harness.ccm.views.graphql.ViewsQueryHelper;
-import io.harness.ccm.views.service.CEViewService;
-import io.harness.ccm.views.service.ViewsBillingService;
-import io.harness.ccm.views.service.impl.ClickHouseViewsBillingServiceImpl;
 
-import com.google.cloud.Timestamp;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import com.google.inject.name.Named;
 import io.leangen.graphql.annotations.GraphQLArgument;
 import io.leangen.graphql.annotations.GraphQLEnvironment;
 import io.leangen.graphql.annotations.GraphQLQuery;
 import io.leangen.graphql.execution.ResolutionEnvironment;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
@@ -64,16 +48,8 @@ import org.jooq.tools.StringUtils;
 @OwnedBy(CE)
 public class PerspectivesQuery {
   @Inject private GraphQLUtils graphQLUtils;
-  @Inject private ViewsBillingService viewsBillingService;
-  @Inject private CEViewService viewService;
-  @Inject private ViewsQueryHelper viewsQueryHelper;
-  @Inject private PerspectiveOverviewStatsHelper perspectiveOverviewStatsHelper;
-  @Inject private PerspectiveTimeSeriesHelper perspectiveTimeSeriesHelper;
-  @Inject private PerspectiveFieldsHelper perspectiveFieldsHelper;
+  @Inject private PerspectiveService perspectiveService;
   @Inject private CCMRbacHelper rbacHelper;
-  @Inject @Named("isClickHouseEnabled") private boolean isClickHouseEnabled;
-  @Inject private ClickHouseViewsBillingServiceImpl clickHouseViewsBillingService;
-  private static final int MAX_LIMIT_VALUE = 10_000;
 
   @GraphQLQuery(name = "perspectiveTrendStats", description = "Trend stats for perspective")
   public PerspectiveTrendStats perspectiveTrendStats(
@@ -83,26 +59,7 @@ public class PerspectivesQuery {
       @GraphQLArgument(name = "isClusterQuery") Boolean isClusterQuery,
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    isClusterQuery = isClusterQuery != null && isClusterQuery;
-    groupBy = groupBy != null ? groupBy : Collections.emptyList();
-
-    ViewQueryParams viewQueryParams = viewsQueryHelper.buildQueryParams(accountId, isClusterQuery);
-
-    // Group by is only needed in case of business mapping
-    if (!viewsQueryHelper.isGroupByBusinessMappingPresent(groupBy)) {
-      viewQueryParams = viewsQueryHelper.buildQueryParamsWithSkipGroupBy(viewQueryParams, true);
-    }
-
-    QLCEViewTrendData trendStatsData =
-        viewsBillingService.getTrendStatsDataNg(filters, groupBy, aggregateFunction, viewQueryParams);
-    return PerspectiveTrendStats.builder()
-        .cost(getStats(trendStatsData.getTotalCost()))
-        .idleCost(getStats(trendStatsData.getIdleCost()))
-        .unallocatedCost(getStats(trendStatsData.getUnallocatedCost()))
-        .systemCost(getStats(trendStatsData.getSystemCost()))
-        .utilizedCost(getStats(trendStatsData.getUtilizedCost()))
-        .efficiencyScoreStats(trendStatsData.getEfficiencyScoreStats())
-        .build();
+    return perspectiveService.perspectiveTrendStats(filters, groupBy, aggregateFunction, isClusterQuery, accountId);
   }
 
   @GraphQLQuery(name = "perspectiveForecastCost", description = "Forecast cost for perspective")
@@ -113,27 +70,7 @@ public class PerspectivesQuery {
       @GraphQLArgument(name = "isClusterQuery") Boolean isClusterQuery,
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    isClusterQuery = isClusterQuery != null && isClusterQuery;
-    groupBy = groupBy != null ? groupBy : Collections.emptyList();
-
-    ViewQueryParams viewQueryParams = viewsQueryHelper.buildQueryParams(accountId, isClusterQuery);
-
-    // Group by is only needed in case of business mapping
-    if (!viewsQueryHelper.isGroupByBusinessMappingPresent(groupBy)) {
-      viewQueryParams = viewsQueryHelper.buildQueryParamsWithSkipGroupBy(viewQueryParams, true);
-    }
-
-    QLCEViewTrendInfo forecastCostData =
-        viewsBillingService.getForecastCostData(filters, groupBy, aggregateFunction, viewQueryParams);
-    return PerspectiveTrendStats.builder()
-        .cost(StatsInfo.builder()
-                  .statsTrend(forecastCostData.getStatsTrend())
-                  .statsLabel(forecastCostData.getStatsLabel())
-                  .statsDescription(forecastCostData.getStatsDescription())
-                  .statsValue(forecastCostData.getStatsValue())
-                  .value(forecastCostData.getValue())
-                  .build())
-        .build();
+    return perspectiveService.perspectiveForecastCost(filters, groupBy, aggregateFunction, isClusterQuery, accountId);
   }
 
   @GraphQLQuery(name = "perspectiveGrid", description = "Table for perspective")
@@ -147,16 +84,8 @@ public class PerspectivesQuery {
       @GraphQLArgument(name = "skipRoundOff") Boolean skipRoundOff,
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    isClusterQuery = isClusterQuery != null && isClusterQuery;
-    skipRoundOff = skipRoundOff != null && skipRoundOff;
-    final int maxLimit = Objects.isNull(limit) ? MAX_LIMIT_VALUE : Integer.min(limit, MAX_LIMIT_VALUE);
-
-    return PerspectiveEntityStatsData.builder()
-        .data(viewsBillingService
-                  .getEntityStatsDataPointsNg(filters, groupBy, aggregateFunction, sortCriteria, maxLimit, offset,
-                      viewsQueryHelper.buildQueryParams(accountId, isClusterQuery, skipRoundOff))
-                  .getData())
-        .build();
+    return perspectiveService.perspectiveGrid(
+        aggregateFunction, filters, groupBy, sortCriteria, limit, offset, isClusterQuery, skipRoundOff, accountId);
   }
 
   @GraphQLQuery(name = "perspectiveFilters", description = "Filter values for perspective")
@@ -169,18 +98,14 @@ public class PerspectivesQuery {
       @GraphQLArgument(name = "isClusterQuery") Boolean isClusterQuery,
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    isClusterQuery = isClusterQuery != null && isClusterQuery;
-
-    return PerspectiveFilterData.builder()
-        .values(viewsBillingService.getFilterValueStatsNg(
-            filters, limit, offset, viewsQueryHelper.buildQueryParams(accountId, isClusterQuery)))
-        .build();
+    return perspectiveService.perspectiveFilters(
+        aggregateFunction, filters, groupBy, sortCriteria, limit, offset, isClusterQuery, accountId);
   }
 
   @GraphQLQuery(name = "perspectiveOverviewStats", description = "Overview stats for perspective")
   public PerspectiveOverviewStatsData perspectiveOverviewStats(@GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    return perspectiveOverviewStatsHelper.fetch(accountId);
+    return perspectiveService.perspectiveOverviewStats(accountId);
   }
 
   @GraphQLQuery(name = "perspectiveTimeSeriesStats", description = "Table for perspective")
@@ -194,73 +119,15 @@ public class PerspectivesQuery {
       @GraphQLArgument(name = "isClusterQuery") Boolean isClusterQuery,
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    final boolean includeOthers = Objects.nonNull(preferences) && Boolean.TRUE.equals(preferences.getIncludeOthers());
-    final boolean includeUnallocatedCost =
-        Objects.nonNull(preferences) && Boolean.TRUE.equals(preferences.getIncludeUnallocatedCost());
-    final int maxLimit = Objects.isNull(limit) ? MAX_LIMIT_VALUE : Integer.min(limit, MAX_LIMIT_VALUE);
-    long timePeriod = perspectiveTimeSeriesHelper.getTimePeriod(groupBy);
-    String conversionField = null;
-    if (viewsBillingService.isDataGroupedByAwsAccount(filters, groupBy)) {
-      conversionField = AWS_ACCOUNT_FIELD;
-    }
-    isClusterQuery = isClusterQuery != null && isClusterQuery;
-
-    String businessMappingId = viewsQueryHelper.getBusinessMappingIdFromGroupBy(groupBy);
-    // If group by business mapping is present, query unified table
-    isClusterQuery = isClusterQuery && businessMappingId == null;
-
-    ViewQueryParams viewQueryParams = viewsQueryHelper.buildQueryParams(accountId, true, false, isClusterQuery, false);
-    Map<String, Map<Timestamp, Double>> sharedCostFromFilters =
-        getSharedCostFromFilters(aggregateFunction, filters, groupBy, sortCriteria, businessMappingId, viewQueryParams);
-    boolean addSharedCostFromGroupBy = false;
-
-    ViewQueryParams viewQueryParamsWithSkipDefaultGroupBy =
-        viewsQueryHelper.buildQueryParams(accountId, true, false, isClusterQuery, false, true);
-
-    PerspectiveTimeSeriesData data;
-    if (isClickHouseEnabled) {
-      data = clickHouseViewsBillingService.getClickHouseTimeSeriesStatsNg(filters, groupBy, aggregateFunction,
-          sortCriteria, includeOthers, maxLimit, viewQueryParams, timePeriod, conversionField, businessMappingId,
-          sharedCostFromFilters, addSharedCostFromGroupBy);
-    } else {
-      data = perspectiveTimeSeriesHelper.fetch(
-          viewsBillingService.getTimeSeriesStatsNg(
-              filters, groupBy, aggregateFunction, sortCriteria, includeOthers, maxLimit, viewQueryParams),
-          timePeriod, conversionField, businessMappingId, accountId, groupBy, sharedCostFromFilters,
-          addSharedCostFromGroupBy);
-    }
-
-    Map<Long, Double> othersTotalCost = Collections.emptyMap();
-    if (includeOthers) {
-      othersTotalCost = viewsBillingService.getOthersTotalCostDataNg(
-          filters, groupBy, Collections.emptyList(), viewQueryParamsWithSkipDefaultGroupBy);
-    }
-
-    Map<Long, Double> unallocatedCost = Collections.emptyMap();
-    if (includeUnallocatedCost) {
-      unallocatedCost = viewsBillingService.getUnallocatedCostDataNg(
-          filters, groupBy, Collections.emptyList(), viewQueryParamsWithSkipDefaultGroupBy);
-    }
-
-    return perspectiveTimeSeriesHelper.postFetch(data, includeOthers, othersTotalCost, unallocatedCost);
-  }
-
-  private Map<String, Map<Timestamp, Double>> getSharedCostFromFilters(List<QLCEViewAggregation> aggregateFunction,
-      List<QLCEViewFilterWrapper> filters, List<QLCEViewGroupBy> groupBy, List<QLCEViewSortCriteria> sortCriteria,
-      String businessMappingId, ViewQueryParams viewQueryParams) {
-    Map<String, Map<Timestamp, Double>> sharedCostFromFilters = new HashMap<>();
-    if (Objects.nonNull(businessMappingId)) {
-      sharedCostFromFilters = viewsBillingService.getSharedCostPerTimestampFromFilters(
-          filters, groupBy, aggregateFunction, sortCriteria, viewQueryParams, viewQueryParams.isSkipRoundOff());
-    }
-    return sharedCostFromFilters;
+    return perspectiveService.perspectiveTimeSeriesStats(
+        aggregateFunction, filters, groupBy, sortCriteria, limit, offset, preferences, isClusterQuery, accountId);
   }
 
   @GraphQLQuery(name = "perspectiveFields", description = "Fields for perspective explorer")
   public PerspectiveFieldsData perspectiveFields(@GraphQLArgument(name = "filters") List<QLCEViewFilterWrapper> filters,
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    return perspectiveFieldsHelper.fetch(accountId, filters);
+    return perspectiveService.perspectiveFields(filters, accountId);
   }
 
   @GraphQLQuery(name = "perspectives", description = "Fetch perspectives for account")
@@ -269,7 +136,7 @@ public class PerspectivesQuery {
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
     if (StringUtils.isEmpty(folderId)) {
-      List<QLCEView> allPerspectives = viewService.getAllViews(accountId, true, sortCriteria);
+      List<QLCEView> allPerspectives = perspectiveService.perspectives(sortCriteria, accountId);
       List<QLCEView> allowedPerspectives = null;
       if (allPerspectives != null) {
         Set<String> allowedFolderIds = rbacHelper.checkFolderIdsGivenPermission(accountId, null, null,
@@ -283,7 +150,7 @@ public class PerspectivesQuery {
     }
     rbacHelper.checkPerspectiveViewPermission(accountId, null, null, folderId);
     return PerspectiveData.builder()
-        .customerViews(viewService.getAllViews(accountId, folderId, true, sortCriteria))
+        .customerViews(perspectiveService.perspectives(folderId, sortCriteria, accountId))
         .build();
   }
 
@@ -293,10 +160,7 @@ public class PerspectivesQuery {
       @GraphQLArgument(name = "isClusterQuery") Boolean isClusterQuery,
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    isClusterQuery = isClusterQuery != null && isClusterQuery;
-
-    return viewsBillingService.getTotalCountForQuery(
-        filters, groupBy, viewsQueryHelper.buildQueryParams(accountId, false, false, isClusterQuery, true));
+    return perspectiveService.perspectiveTotalCount(filters, groupBy, isClusterQuery, accountId);
   }
 
   @GraphQLQuery(name = "workloadLabels", description = "Labels for workloads")
@@ -304,19 +168,6 @@ public class PerspectivesQuery {
       @GraphQLArgument(name = "filters") List<QLCEViewFilterWrapper> filters,
       @GraphQLEnvironment final ResolutionEnvironment env) {
     final String accountId = graphQLUtils.getAccountIdentifier(env);
-    return viewsBillingService.getLabelsForWorkloads(accountId, workloads, filters);
-  }
-
-  private StatsInfo getStats(QLCEViewTrendInfo trendInfo) {
-    if (trendInfo == null) {
-      return null;
-    }
-    return StatsInfo.builder()
-        .statsTrend(trendInfo.getStatsTrend())
-        .statsLabel(trendInfo.getStatsLabel())
-        .statsDescription(trendInfo.getStatsDescription())
-        .statsValue(trendInfo.getStatsValue())
-        .value(trendInfo.getValue())
-        .build();
+    return perspectiveService.workloadLabels(workloads, filters, accountId);
   }
 }
