+++ batch-processing/service/batch-processing-config.yml
@@ -218,20 +218,3 @@ azureVmPricingConfig:
   baseUrl: https://prices.azure.com/api/retail/
   connectTimeOutSeconds: 15
   readTimeOutSeconds: 15
-
-accessControlClient:
-  enableAccessControl: true
-  accessControlServiceConfig:
-    baseUrl: http://localhost:9006
-    connectTimeOutSeconds: 15
-    readTimeOutSeconds: 15
-  accessControlServiceSecret: IC04LYMBf1lDP5oeY4hupxd4HJhLmN6azUku3xEbeE3SUx5G3ZYzhbiwVtK4i7AmqyU9OZkwB4v8E9qM
-
-accessControlAdminClient:
-  accessControlServiceConfig:
-    baseUrl: http://localhost:9006/api/
-    connectTimeOutSeconds: 15
-    readTimeOutSeconds: 15
-    enableHttpLogging: false
-  accessControlServiceSecret: IC04LYMBf1lDP5oeY4hupxd4HJhLmN6azUku3xEbeE3SUx5G3ZYzhbiwVtK4i7AmqyU9OZkwB4v8E9qM
-  mockAccessControlService: true
