+++ batch-processing/service/src/main/java/io/harness/batch/processing/config/BatchProcessingModule.java
@@ -10,8 +10,6 @@ package io.harness.batch.processing.config;
 import static io.harness.authorization.AuthorizationServiceHeader.BATCH_PROCESSING;
 import static io.harness.authorization.AuthorizationServiceHeader.MANAGER;
 
-import io.harness.AccessControlClientModule;
-import io.harness.accesscontrol.AccessControlAdminClientModule;
 import io.harness.account.AccountClient;
 import io.harness.annotations.retry.MethodExecutionHelper;
 import io.harness.annotations.retry.RetryOnException;
@@ -60,8 +58,6 @@ import io.harness.ccm.jira.CCMJiraHelper;
 import io.harness.ccm.jira.CCMJiraHelperImpl;
 import io.harness.ccm.msp.service.impl.MarginDetailsServiceImpl;
 import io.harness.ccm.msp.service.intf.MarginDetailsService;
-import io.harness.ccm.rbac.CCMRbacHelper;
-import io.harness.ccm.rbac.CCMRbacHelperImpl;
 import io.harness.ccm.service.impl.AWSOrganizationHelperServiceImpl;
 import io.harness.ccm.service.intf.AWSOrganizationHelperService;
 import io.harness.ccm.views.businessmapping.service.impl.BusinessMappingHistoryServiceImpl;
@@ -213,7 +209,6 @@ public class BatchProcessingModule extends AbstractModule {
     bind(AnomalyService.class).to(AnomalyServiceImpl.class);
     bind(MarginDetailsService.class).to(MarginDetailsServiceImpl.class);
     bind(ManagedAccountDataService.class).to(ManagedAccountDataServiceImpl.class);
-    bind(CCMRbacHelper.class).to(CCMRbacHelperImpl.class);
     install(new ConnectorResourceClientModule(batchMainConfig.getNgManagerServiceHttpClientConfig(),
         batchMainConfig.getNgManagerServiceSecret(), BATCH_PROCESSING.getServiceId(), ClientMode.PRIVILEGED));
     install(new InstanceNGResourceClientModule(batchMainConfig.getNgManagerServiceHttpClientConfig(),
@@ -224,10 +219,6 @@ public class BatchProcessingModule extends AbstractModule {
         batchMainConfig.getCeNgServiceSecret(), BATCH_PROCESSING.getServiceId(), ClientMode.PRIVILEGED));
     install(new SecretNGManagerClientModule(batchMainConfig.getNgManagerServiceHttpClientConfig(),
         batchMainConfig.getNgManagerServiceSecret(), BATCH_PROCESSING.getServiceId()));
-    install(AccessControlClientModule.getInstance(
-        batchMainConfig.getAccessControlClientConfiguration(), BATCH_PROCESSING.getServiceId()));
-    install(new AccessControlAdminClientModule(
-        batchMainConfig.getAccessControlAdminClientConfiguration(), BATCH_PROCESSING.getServiceId()));
     install(new AzureVmPricingClientModule(batchMainConfig.getAzureVmPricingConfig()));
     install(new AbstractTelemetryModule() {
       @Override
